+++ tests/testthat/test-diagnostics.R
@@ -13,11 +13,11 @@ test_that("Fitted model extracts diagnostics", {
       "low_case_count_flag"
     ),
     value = c(
-      0.6990423,
+      0.9145338,
       0.0000000,
       0.0000000,
       0.0000000,
-      0.1391304,
+      0.1415929,
       16.0000000,
       1.0000000,
       1.0000000